+++ spring-code-for-deploy/src/main/java/com/example/seminar/dto/request/member/MemberProfileUpdateRequest.java
@@ -5,7 +5,7 @@ import lombok.Data;
 
 @Data
 public class MemberProfileUpdateRequest {
-    private int generation;
+    private short generation;
     private Part part;
 }
