+++ spring-code-for-deploy/src/main/java/com/example/seminar/repository/CategoryJpaRepository.java
@@ -3,5 +3,5 @@ package com.example.seminar.repository;
 import com.example.seminar.domain.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryJpaRepository extends JpaRepository<Category, Short> {
+public interface CategoryJpaRepository extends JpaRepository<Category, Integer> {
 }
