+++ spring-code-for-deploy/src/main/java/com/example/seminar/service/category/CategoryService.java
@@ -1,7 +1,6 @@
 package com.example.seminar.service.category;
 
 
-import com.example.seminar.common.exception.BusinessException;
 import com.example.seminar.domain.Category;
 import com.example.seminar.domain.CategoryId;
 import com.example.seminar.dto.response.category.CategoryResponse;
@@ -19,11 +18,11 @@ public class CategoryService {
     private final CategoryJpaRepository categoryJpaRepository;
 
     public Category getByCategoryId(CategoryId categoryId) {
-        return categoryJpaRepository.findById(Short.valueOf(categoryId.getCategoryId())).orElseThrow(
+        return categoryJpaRepository.findById(Integer.valueOf(categoryId.getCategoryId())).orElseThrow(
                 () -> new EntityNotFoundException("해당하는 카테고리가 없습니다."));
     }
 
-    public CategoryResponse getById(Short id) {
+    public CategoryResponse getById(int id) {
         return CategoryResponse.of(categoryJpaRepository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException("해당하는 카테고리가 없습니다.")));
     }
